// corporate-website/pages/technologies/technology-type.ts
/**
 * Represents a technology developed by Corporate Inc.
 */
export type TechnologyType = {
  /**
   * Unique identifier for the technology
   */
  id: string;
  
  /**
   * Title of the technology
   */
  title: string;
  
  /**
   * Detailed description of the technology
   */
  description: string;
  
  /**
   * URL to an image representing the technology
   */
  imageUrl: string;
  
  /**
   * Category the technology belongs to
   */
  category: string;
};